datatype change

+++ ArmsServices/DataServices/Operations/DestinationFeasibilityCheckerService.cs
@@ -124,19 +124,19 @@ namespace ArmsServices.DataServices
                 ID = dr.GetInt32("ID"),
                 Content = new ContentModel
                 {
-                    ContentID = dr.GetInt16("ContentID"),
+                    ContentID = dr.GetByte("ContentID"),
                     ContentName = dr.GetString("ContentName"),
                 },
                 BodyType = dr.GetString("BodyType"),
                 TruckType = new TruckTypeModel
                 {
-                    TruckTypeID = dr.GetInt16("TruckTypeID"),
+                    TruckTypeID = dr.GetByte("TruckTypeID"),
                     TruckType = dr.GetString("TruckType"),
                     wheels = dr.GetByte("wheels"),
                     BSType = dr.GetString("BSType"),
                 },
-                SystemKM = dr.GetInt32("SystemKM"),
-                RunKM = dr.GetInt32("RunKM"),
+                SystemKM = dr.GetDecimal("SystemKM"),
+                RunKM = dr.GetDecimal("RunKM"),
                 StandardDays = dr.GetDecimal("StandardDays"),
                 FuelLitre = dr.GetDecimal("FuelLitre"),
                 StandardMileage = dr.GetDecimal("StandardMileage"),
